+++ src/query/plan/rule_based_planner.cpp
@@ -359,7 +359,7 @@ class ReturnBodyContext : public HierarchicalTreeVisitor {
   VISIT_BINARY_OPERATOR(LessEqualOperator)
   VISIT_BINARY_OPERATOR(GreaterEqualOperator)
   VISIT_BINARY_OPERATOR(InListOperator)
-  VISIT_BINARY_OPERATOR(ListMapIndexingOperator)
+  VISIT_BINARY_OPERATOR(SubscriptOperator)
 
 #undef VISIT_BINARY_OPERATOR